+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* add missing `BLen (GenericallyNonSum a)` instance
+
 ## 0.8.0 (2024-04-13)
 * add missing `And` predicate combinators instances (`PutC`, `GetC`)
 * add `Type.Derived.NullTermPadded` (type synonym over `And`)